structured logging: replacing Fatal/Exit/etc. without loss of flushing
When structured logging called for functions like Fatal and Exit to be avoided
in favor of klog.ErrorS + os.Exit, it was overlooked that this replacement
skips flushing.

We could ask developers to use klog.ErrorS + klog.Flush + os.Exit, but that is
still not quite the same because klog.Flush could get stuck or take too long.
Therefore klog now supports flushing + exiting with a dedicated function that
can be combined with some other, arbitrary logging function.

Allowing applications to change the default timeout and how specifically the
application terminates may be useful, so both values are variables that are
exported and users of klog are allowed to change.
